[REF] core: organize base's tests to better match models
