+++ odoo/addons/base/tests/__init__.py
@@ -1,40 +1,36 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import common
-from . import test_base
-from . import test_format_address_mixin
-from . import test_groups
 from . import test_avatar_mixin
 from . import test_ir_actions
+from . import test_ir_actions_report
 from . import test_ir_asset
 from . import test_ir_attachment
+from . import test_ir_config_parameter
 from . import test_ir_cron
+from . import test_ir_default
+from . import test_ir_embedded_actions
 from . import test_ir_filters
 from . import test_ir_http
 from . import test_ir_mail_server
 from . import test_ir_mail_server_smtpd
 from . import test_ir_model
 from . import test_ir_module
+from . import test_ir_qweb
+from . import test_ir_qweb_fields
 from . import test_ir_sequence
-from . import test_ir_sequence_date_range
-from . import test_ir_embedded_actions
-from . import test_ir_default
-from . import test_menu
-from . import test_qweb_field
-from . import test_qweb
-from . import test_res_config
-from . import test_res_lang
-from . import test_split_table
-from . import test_user_has_group
-from . import test_views
+from . import test_ir_ui_menu
+from . import test_ir_ui_view
 from . import test_res_company
-from . import test_res_currency
+from . import test_res_config
 from . import test_res_country
+from . import test_res_currency
+from . import test_res_groups
+from . import test_res_groups_data
+from . import test_res_lang
 from . import test_res_partner
+from . import test_res_partner_addresses
 from . import test_res_partner_bank
-from . import test_res_partner_merge
 from . import test_res_users
-from . import test_reports
-from . import test_form_create
-from . import test_config_parameter
-from . import test_ir_module_category
+from . import test_res_users_has_group
+from . import test_wizard_base_partner_merge
