+++ odoo/addons/base/tests/test_ir_config_parameter.py
