+++ odoo/addons/base/tests/test_ir_module.py
@@ -1,3 +1,4 @@
+from odoo.exceptions import ValidationError
 from odoo.tests.common import tagged, TransactionCase
 from odoo.tools import mute_logger
 
@@ -48,3 +49,22 @@ class IrModuleCase(TransactionCase):
         self.assertEqual(module.state, "installed")
         module.module_uninstall()
         self.assertEqual(module.state, "uninstalled")
+
+
+@tagged('at_install', '-post_install')
+class TestModuleCategory(TransactionCase):
+
+    def test_parent_circular_dependencies(self):
+        Cats = self.env['ir.module.category']
+
+        def create(name, **kw):
+            return Cats.create(dict(kw, name=name))
+
+        category_a = create('A', parent_id=False)
+        category_b = create('B', parent_id=category_a.id)
+        category_c = create('C', parent_id=category_b.id)
+
+        with self.assertRaises(ValidationError):
+            category_a.write({'parent_id': category_c.id})
+        with self.assertRaises(ValidationError):
+            category_b.write({'parent_id': category_b.id})
