+++ odoo/addons/base/tests/test_ir_qweb_fields.py
