+++ odoo/addons/base/tests/test_ir_ui_menu.py
