+++ odoo/addons/base/tests/test_res_groups_data.py
