+++ odoo/addons/base/tests/test_res_users_has_group.py
