+++ odoo/addons/base/tests/test_wizard_base_partner_merge.py
