+++ odoo/addons/test_orm/tests/test_views.py
@@ -1,7 +1,7 @@
 from lxml import etree
 
 from odoo.tests import tagged, common
-from odoo.addons.base.tests.test_views import ViewCase
+from odoo.addons.base.tests.test_ir_ui_view import ViewCase
 
 
 @tagged('at_install', '-post_install')  # LEGACY at_install
